✨(api) open an API endpoint to retrieve a single course run
As we implement the enrollment button in the frontend, we will need
to be able to determine what the button should display and if it
should be displayed at all.

Part of these determinations is to get the course run object along
with its state property. Hence our opening up of the API to retrieve
a course run (which is what will be needed for this use-case).