✨(enrollment) add a filter to check if a course has managed enrollments
We have all the logic in place to manage enrollments for courses through
a frontend component and API endpoints it can connect to. The last
missing piece is incorporating the frontend component in the template,
for the course runs (and Richie installs) that need it.

Rather than force everything to behave the same way through a Richie-wide
setting, we've chosen to add a filter that enables each course run
template to check for a matching LMS. That way, courses with managed
enrollments can coexist on a single Richie instance with courses that
simply provide a link to an external LMS.

For Richie instances with no managed enrollments, the filter will simply
always return False and default to the "link to external LMS" approach.