✨(courses) integrate course run enrollment in course views
We need to integrate course run enrollments, through our new
<CourseRunEnrollment /> component, on the course detail view and the
course run detail view. We want to make enrollments seamless for users
but still give them links to the course run views even when there is
an enrollment button.

We're using our `has_connected_lms()` filter to determine if
a course run (be it on its own view or on the course view) should display
the resource link "as is" or manage enrollments.